Refactor api.py, create models and controller modules
